fookin uninitialised value in deathcheck

+++ utils.c
@@ -73,7 +73,7 @@ void	message(int ph_id, t_data *data, char *s)
 	pthread_mutex_lock(&(data->message));
 	if (data->dead == 0)
 	{
-		printf("\033[00m%li ms \033 ", get_the_time() -  data->first_time);
+		printf("\033[00m%lli ms \033 ", get_the_time() - (long)data->first_time);
 		printf("\033[00m%i \033 ", (ph_id + 1));
 		printf("%s", s);
 	}
